third_party/blink/web_tests/http/tests/security/script-onerror-crossorigin-cors-expected.txt

CONSOLE ERROR: Uncaught SomeError
The test passes if 'window.onerror' gets unsanitized information about an exception thrown in a script loaded with a 'crossorigin' attribute, and delivered with valid CORS headers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS /SomeError/.test(msg) is true
PASS url is "http://localhost:8000/security/resources/cors-script.php?fail=true&cors=true"
PASS line is 1
PASS column is 1
PASS window.errorObject is not null
PASS successfullyParsed is true

TEST COMPLETE
